+++ src/components/links-bread-crumbs/messages.ts
@@ -1,8 +1,3 @@
 export const messages = {
   homeLink: 'Home',
-  aboutLink: 'About',
-  notFoundLink: '404 Not Found',
-  loginLink: 'Login',
-  registerLink: 'Sign up',
-  cartLink: 'Cart',
 };
