--- src/sources/constants/page-title/messages.ts
@@ -1,8 +1,0 @@
-export const messages = {
-  mainPageTitle: 'Main Page',
-  aboutPageTitle: 'About Us',
-  notFoundPageTitle: '404 Not Found',
-  loginPageTitle: 'Sign In',
-  registerPageTitle: 'Sign Up',
-  cartPageTitle: 'Cart',
-};
