+++ src/sources/messages.ts
@@ -1,0 +1,8 @@
+export const messages = {
+  mainPageTitle: 'Main Page',
+  aboutPageTitle: 'About Us',
+  notFoundPageTitle: '404 Not Found',
+  loginPageTitle: 'Sign In',
+  registerPageTitle: 'Sign Up',
+  cartPageTitle: 'Cart',
+};
